7980 Fix two config file issues
